Bringing latest changes to Main branch (#9)
* Some test python scripts

* 07.10.2023_[Compiling, Not Working, Generating Timecourses]_C++ CUDA Code

* some test scripts

* MAJOR: Fully working first version of GPU-pRF analysis, results anaylsis scripts etc.

* changed the handling of configuration file

* Some Sandbox test scripts

* Corrected result verification script (mrVista and opRF)

* minor path change in the cuda kernel call in old script

* MAJOR: Fully working class of HPC opRF and a test script for it

* just setting sigma to single value for old script

* Bugfix: Use of "nanargmax( )" for grid-fit

* Bugfix: Stimulus resampling ratio calculation

* Important: New  R2 calculation (with both possibilities as e =yts and e = f(x))

* Some new test scripts

* Bugfix in R2 calculation. Managing NaN

* MAJOR: Performance enahanced computation of R2, and Gaussian Curves kernel to pass (ux, uy, sigma) arrays

* Batch based GPU processing

* Some Sandbox and Results-Verification scripts

* Major: Performance test, Frequency filtering (Test), minor modificaitons scripts

* Add a dummy config for CEPH data

* some more test scripts

* Required Soluion file for the added scripts

* Important: changed config file (for results field), stimulus on gpu function

* some further test scripts

* Major Bugfix: correcting Sigma calculation in CUDA kernel

* Minor comment

* Minor renaming

* Bugfix: fixing prime-star signals

* Super MAJOR: Quiver plots, gradients test, refine-fit debug code

* Major: true multi-gpu with batching, working (dgx tested)

* Super MAJOR: v1 of Concatenated runs, tested only with same stimulus and same data

* Finding all available GPUs

* Major: error worsening resolved, debugging code, line-line intersection, singleton search-space, adapted analysis scripts, gradients analysis scripts, ECCENTRICITY analysis script (005-gpu-pRF-analysis)

* some scripts

* Load from PTX - Sample  program

* sample sandbox scripts

* GEM-pRF - First working version. Still the verification of Errors improvement needs to be done after refinement.

* dummy file for Bids Gem analysis

* some file and foldr structures

* MAJOR: Fully working Bids compatible GEM, with a separate RUN script

* Bugfix: Correct the batching of GPU signals computation

* script to plot NumVoxels vs ExecutionTimes plot

* MAJOR: Concatenation run (generic, tested)

* Tools: GUI app to compare GEM vs Vista vs Nordic (Seaborn plots)

* Minor changes

* MAJOR BUGFIX: Indexing issue in case of Multi-GPU system

* MAJOR: Validation of multi-dimensional points for pRFSpace

* Minor: Removing dependencies from OpRF

* Bugfix: single y-signal in a batch

* Test multiprocessing and multithreading functions in PRFSpace (identical results)

* minor adjutment

* Cleaned up PRFSpace file

+++ codebase-cpp/fmri-hpc/utils/arrays/Linspace.cpp
@@ -1,8 +1,8 @@
 #include "Linspace.h"
 
-Linspace::Linspace(float start, float end, int numPoints) : _start(start), _end(end), _numPoints(numPoints)
+Linspace::Linspace(double start, double end, int numPoints) : _start(start), _end(end), _numPoints(numPoints)
 {    
-    _resultArr = new float[_numPoints];
+    _resultArr = new double[_numPoints];
 
     if (numPoints > 1)
         _step = (_end - _start) / (_numPoints - 1);
@@ -10,9 +10,9 @@ Linspace::Linspace(float start, float end, int numPoints) : _start(start), _end(
         _step = 1;
 }
 
-std::vector<float> Linspace::Generate()
+std::vector<double> Linspace::Generate()
 {
-    std::vector<float> result;
+    std::vector<double> result;
     result.reserve(_numPoints);
 
     for (int i = 0; i < _numPoints; ++i)
@@ -23,11 +23,11 @@ std::vector<float> Linspace::Generate()
     return result;
 }
 
-float* Linspace::GenerateLinspaceArr()
+double* Linspace::GenerateLinspaceArr()
 {
     for (int i = 0; i < _numPoints; ++i)
     {
-        _resultArr[i] = static_cast<float>(_start + i * _step);
+        _resultArr[i] = static_cast<double>(_start + i * _step);
     }
 
     return _resultArr;
