+++ codebase-cpp/fmri-hpc/utils/arrays/Linspace.h
@@ -5,15 +5,15 @@
 class Linspace
 {
 public:
-    Linspace(float start, float end, int numPoints);
-    std::vector<float> Generate();
-    float* GenerateLinspaceArr();
+    Linspace(double start, double end, int numPoints);
+    std::vector<double> Generate();
+    double* GenerateLinspaceArr();
     ~Linspace();
 
 private:
-    float _start;
-    float _end;
+    double _start;
+    double _end;
     int _numPoints;
-    float _step;
-    float* _resultArr;
+    double _step;
+    double* _resultArr;
 };
