+++ codebase-cpp/fmri-hpc/utils/arrays/Meshgrid.h
@@ -4,18 +4,18 @@
 class Meshgrid 
 {
 public:
-    Meshgrid(const std::vector<float>& x, const std::vector<float>& y);
-    std::vector<std::vector<float>> X();
-    std::vector<std::vector<float>> Y();
-    float* XMeshGridArr();
-    float* YMeshGridArr();
+    Meshgrid(const std::vector<double>& x, const std::vector<double>& y);
+    std::vector<std::vector<double>> X();
+    std::vector<std::vector<double>> Y();
+    double* XMeshGridArr();
+    double* YMeshGridArr();
     ~Meshgrid();
 
 private:
-    std::vector<float> _x;
-    std::vector<float> _y;
+    std::vector<double> _x;
+    std::vector<double> _y;
     int _numX;
     int _numY;
-    float* _xMeshGridArr;
-    float* _yMeshGridArr;
+    double* _xMeshGridArr;
+    double* _yMeshGridArr;
 };
